// src/components/header-short-component/HeaderShortComponent.interface.d.ts
import { GlobalVariantExtended, GlobalVariants } from '../../interfaces/global-variants';

export interface HeaderShortComponentProps {
    variant?: GlobalVariants | GlobalVariantExtended;
    showArrow?: boolean;
    tagline?: string;
    headline?: string;
    className?: string;
}
//# sourceMappingURL=HeaderShortComponent.interface.d.ts.map